% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseflow.R
\name{baseflow}
\alias{baseflow}
\title{Calculation of baseflow based on smoothed minima method}
\usage{
baseflow(x, tp.factor = 0.9, block.len = 5)
}
\arguments{
\item{x}{numeric vector with observed streamflow, normally given in daily values}

\item{tp.factor}{numeric vector, turning point factor}

\item{block.len}{numeric vector, block size of filter to set flow minima}
}
\value{
Baseflow series.
}
\description{
Original \code{baseflow} function from the \code{lfstat} package.
}
\examples{
\dontrun{
baseflow(q_data$q_obs, block.len = 10)
}
}
\references{
lfstat: Calculation of Low Flow Statistics for Daily Stream Flow Data
D. Koffler, T. Gauster, T., and G. Laaha (https://CRAN.R-project.org/package=lfstat)
}
