% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfi_n.R
\name{dfi_n}
\alias{dfi_n}
\title{Calculation of DFI for block length of N}
\usage{
dfi_n(q, n = 5)
}
\arguments{
\item{q}{streamflow series}

\item{n}{block length}

\item{...}{other parameters passed to function}
}
\value{

}
\description{
Calculation of DFI for block length of N
}
\examples{
dfi_n(q_obs, n = 2:5)
}
