% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_bps.R
\name{find_bp}
\alias{find_bp}
\title{Find breakpoints in DFI curve}
\usage{
find_bp(
  dfi,
  n_bp = 3,
  min_gp_gap = 5,
  filter_min = 0,
  filter_max = 120,
  dfi_check = TRUE,
  print = FALSE,
  plotting = FALSE
)
}
\arguments{
\item{dfi}{numeric, DFI values between 1 and 0}

\item{n_bp}{numeric, number of breakpoints}

\item{min_gp_gap}{numeric, smallest interval between two breakpoints}

\item{filter_min}{numeric, \code{filter_min}+1 is minimum potential breakpoint estimate}

\item{filter_max}{numeric, \code{filter_max}-1 is maximum potential breakpoint estimate}

\item{dfi_check}{logical, if \code{TRUE} DFI values are converted to be continuously descreasing with cummin()}

\item{print}{logical, if \code{TRUE} calculation of breakpoints are printed}

\item{plotting}{logical, if \code{TRUE} the DFI curve and piecewise linear segments are plotted}
}
\value{
Returns a list.
\item{breakpoints}{estimates for the n breakpoints with names \code{bp_x}}
\item{bias}{value of the objective function \verb{OF = 1/2 RMSE + 1/2 MAE}}
\item{rel_contr}{Relative streamflow contributions between \code{filter_min}, the breakpoints and \code{filter_max},
e.g. 2 breakpoints lead to 3 relative contributions, 3 breakpoints lead to 4 relative contributions}
}
\description{
Find breakpoints in DFI curve
}
\examples{
# use dfi_example as an DFI vector with 121 values
find_bp(dfi_example, n_bp = 2, filter_max = 90, dfi_check = FALSE)
}
\references{
Stoelzle, M., Schuetz, T., Weiler, M., Stahl, K., & Tallaksen, L. M. (2020). Beyond binary baseflow separation: a delayed-flow index for multiple streamflow contributions. Hydrology and Earth System Sciences, 24(2), 849-867.
}
