% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_bps.R
\name{find_bps}
\alias{find_bps}
\title{Find breakpoints in DFI curve}
\usage{
find_bps(
  dfi,
  n_bp = 2,
  bp_mingap = 5,
  bp_min = 0,
  bp_max = length(dfi) - 1,
  nmax = length(dfi),
  of_weights = c(0.5, 0.5),
  desc = TRUE,
  print = FALSE,
  plotting = FALSE
)
}
\arguments{
\item{dfi}{numeric, a vector with the DFI values between 1 and 0, dummy data \link{dfi_example} can be used.}

\item{n_bp}{numeric, How many breakpoints (1, 2, 3) should be estimated (default = 2)?}

\item{bp_mingap}{numeric, smallest interval between two breakpoints (default = 5).}

\item{bp_min}{numeric, \code{bp_min}+1 is minimum allowed breakpoint position (default = 0).}

\item{bp_max}{numeric, \code{bp_max}-1 is maximum allowed breakpoint position.}

\item{nmax}{numeric parameter to truncate the tailing of the CDC (i.e. the length of \code{dfi} vector).
Note, \code{nmax} should not be confused with \code{bp_max} (i.e. maximum breakpoint position).}

\item{of_weights}{vector with two elements, first is weight of the RMSE, second is weight of the MAE.
Default is c(0.5, 0.5), i.e. equal weights. Sum of vector must be 1. To switch of one measure use
\code{c(1, 0)} or \code{c(0,1)}. With the weighting more or less focus could be given to the upper parts
of the DFI curve (i.e. when block length \code{n} for separation is between 1 and 10).}

\item{desc}{logical, if \code{TRUE} DFI values are converted to be monotonically decreasing with \code{\link[=cummin]{cummin()}}}

\item{print}{logical, if \code{TRUE} best breakpoint estimates during calculation are printed (debug mode)}

\item{plotting}{logical, if \code{TRUE} the DFI curve and piecewise linear segments are plotted with \code{plot()}}
}
\value{
Returns a list.
\item{breakpoints}{estimates for the n breakpoints with names \code{bp_x}}
\item{bias}{value of the objective function
\verb{OF = 1/2 RMSE + 1/2 MAE},
where OF is to be minimized.}
\item{rel_contr}{Relative streamflow contributions between \code{bp_min}, the breakpoints and \code{bp_max},
e.g. 2 breakpoints lead to 3 relative contributions, 3 breakpoints lead to 4 relative contributions.
The first \code{rel_contr}-avlue is the fastest contribution to streamflow, the last \code{rel_contr}-value
is the slowest contribution to streamflow.}
}
\description{
Estimates n breakpoints in the DFI curve.
DFI values are compared against n+1 piecewise linear segments and the minimization of the
corresponding residuals. Several parameters can be customized.
If \code{q_obs} is given in daily temporal resolution then also \code{n} can be interpreted as delay in days
for different contributions (default).
}
\examples{
# use dfi_example as an DFI vector with 121 values
find_bps(dfi_example, n_bp = 2, bp_max = 90, plotting = TRUE)
}
\references{
Stoelzle, M., Schuetz, T., Weiler, M., Stahl, K., & Tallaksen, L. M. (2020). Beyond binary baseflow separation: a delayed-flow index for multiple streamflow contributions. Hydrology and Earth System Sciences, 24(2), 849-867.
}
