% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mae.R
\name{mae}
\alias{mae}
\title{Mean Absolute Error}
\usage{
mae(sim, obs)
}
\arguments{
\item{sim}{numeric, simulated values}

\item{obs}{numeric, observed values}
}
\value{
Mean absolute error between sim and obs.
}
\description{
Mean Absolute Error
}
\examples{
mae(c(1:3), (2:4))
}
