% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_example.R
\docType{data}
\name{q_data}
\alias{q_data}
\title{Example streamflow data}
\format{
data.frame
}
\usage{
q_data
}
\description{
18 years of daily streamflow from river Dreisam (Germany). Values are m3/s.
Data provided by the LUBW Baden-Württemberg (Landesanstalt fuer Umwelt,
State Institute of Environment, https://www.lubw.baden-wuerttemberg.de/startseite)
}
\examples{
head(q_data, 5)

tail(q_data, 5)
}
\keyword{datasets}
