% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{Root Mean Square Error}
\usage{
rmse(sim, obs)
}
\arguments{
\item{sim}{numeric with simulated values}

\item{obs}{numeric with observed values}
}
\value{
Root mean square error (rmse) between sim and obs.
}
\description{
Root Mean Square Error
}
\examples{
rmse(c(1:3), c(1, 3, 5))
}
