% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_min_na.R
\name{which.min.na}
\alias{which.min.na}
\title{Helper Function for which.min(x)}
\usage{
which.min.na(x)
}
\arguments{
\item{x}{numeric verctor}
}
\value{
Position of minimum value in the vector
}
\description{
Calculation of minimum position in a vector with NA values.
}
\examples{
which.min.na(c(2,3,4,1,NA,6)) # returns 4
}
