% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_bps.R
\name{find_bp}
\alias{find_bp}
\title{Find breakpoints in DFI curve}
\usage{
find_bp(
  dfi,
  n_bp = 3,
  min_gp_gap = 5,
  filter_min = 0,
  filter_max = 120,
  dfi_check = TRUE,
  print = FALSE,
  plotting = FALSE
)
}
\arguments{
\item{dfi}{numeric, DFI values between 1 and 0}

\item{n_bp}{numeric, number of breakpoints}

\item{min_gp_gap}{numeric, smallest interval between two breakpoints}

\item{filter_min}{numeric, filter_max+1 is minimum potential breakpoint estimate}

\item{filter_max}{numeric, filter_max-1 is maximum potential breakpoint estimate}

\item{dfi_check}{logical, if TRUE DFI values are converted to be continuously descreasing with cummin()}

\item{print}{logical, if TRUE calculation of breakpoints are printed}

\item{plotting}{logical, if TRUE the DFI curve and piecewise linear segments are plotted}
}
\value{
Returns a list.
\item{breakpoints}{estimates for the n breakpoints with names "bp_x"}
\item{bias}{value of the objective function OF = 1/2 RMSE + 1/2 MAE"}
\item{rel_contr}{Relative streamflow contributions between \code{filter_min}, the breakpoints and \code{filter_max}}
}
\description{
Find breakpoints in DFI curve
}
\examples{
# use dfi_example as an DFI vector with 121 values
find_bp(dfi_example, n_bp = 2, filter_max = 90, dfi_check = FALSE)
}
