% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_bps.R
\name{find_bp}
\alias{find_bp}
\title{Find breakpoints in DFI curve}
\usage{
find_bp(
  dfi,
  n_bp = 2,
  min_gp_gap = 5,
  filter_min = 0,
  filter_max = length(dfi) - 1,
  dfi_check = TRUE,
  print = FALSE,
  plotting = FALSE
)
}
\arguments{
\item{dfi}{numeric, a vector with the DFI values between 1 and 0, dummy data \link{dfi_example} can be used.}

\item{n_bp}{numeric, How many breakpoints (1, 2, 3) should be estimated? Default = 2}

\item{min_gp_gap}{numeric, smallest interval between two breakpoints. Default = 5 (days)}

\item{filter_min}{numeric, \code{filter_min}+1 is minimum potential breakpoint estimate}

\item{filter_max}{numeric, \code{filter_max}-1 is maximum potential breakpoint estimate}

\item{dfi_check}{logical, if \code{TRUE} DFI values are converted to be monotonically descreasing with \code{\link[=cummin]{cummin()}}}

\item{print}{logical, if \code{TRUE} best breakpoint estimates during calculation are printed (debug mode)}

\item{plotting}{logical, if \code{TRUE} the DFI curve and piecewise linear segments are plotted with \code{plot()}}
}
\value{
Returns a list.
\item{breakpoints}{estimates for the n breakpoints with names \code{bp_x}}
\item{bias}{value of the objective function \verb{OF = 1/2 RMSE + 1/2 MAE}}
\item{rel_contr}{Relative streamflow contributions between \code{filter_min}, the breakpoints and \code{filter_max},
e.g. 2 breakpoints lead to 3 relative contributions, 3 breakpoints lead to 4 relative contributions}
}
\description{
Estimates n breakpoints in the DFI curve.
DFI values are compared against n+1 piecewise linear segments and the minimization of the
corresponding residuals. Several parameters can be tweaked.
}
\examples{
# use dfi_example as an DFI vector with 121 values
find_bp(dfi_example, n_bp = 2, filter_max = 90, dfi_check = FALSE)
}
\references{
Stoelzle, M., Schuetz, T., Weiler, M., Stahl, K., & Tallaksen, L. M. (2020). Beyond binary baseflow separation: a delayed-flow index for multiple streamflow contributions. Hydrology and Earth System Sciences, 24(2), 849-867.
}
