% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nmax.R
\name{find_nmax}
\alias{find_nmax}
\title{Find nmax based on low flow thresholds}
\usage{
find_nmax(df, n = 1:180)
}
\arguments{
\item{df}{data.frame with two columns, first column must be Date, second must be streamflow
data.}

\item{n}{numeric vector}
}
\value{
A list.
}
\description{
Find nmax based on low flow thresholds
}
\examples{
find_nmax(q_data)
}
