% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_example.R
\docType{data}
\name{q_data}
\alias{q_data}
\title{Example streamflow data}
\format{
data.frame
}
\usage{
q_data
}
\description{
18 years of daily streamflow from river Dreisam (Germany). Values are m3/s.
}
\examples{
head(q_data,25)
tail(q_data,25)
}
\keyword{datasets}
